# [ENHANCEMENT] [MER-3391] update codebase and build infrastructure for elixir 1 17 0 and erlang otp 27