--- a/.formatter.exs
+++ b/.formatter.exs
@@ -4,8 +4,7 @@
   inputs: [
     "*.{heex,ex,exs}",
     "priv/*/seeds.exs",
-    "{config,lib,test}/**/*.{heex,ex,exs}",
-    "storybook/**/*.exs"
+    "{config,lib,test}/**/*.{heex,ex,exs}"
   ],
   subdirectories: ["priv/*/migrations"]
 ]