--- a/.github/workflows/build.yml
+++ b/.github/workflows/build.yml
@@ -52,8 +52,8 @@ jobs:
       - name: 🧪 Setup elixir
         uses: erlef/setup-elixir@v1
         with:
-          elixir-version: 1.15.5 # Define the elixir version [required]
-          otp-version: 26.1 # Define the OTP version [required]
+          elixir-version: 1.17.2 # Define the elixir version [required]
+          otp-version: 27.0.1 # Define the OTP version [required]
 
       - name: ⬇️ Install elixir dependencies
         run: mix deps.get
@@ -123,8 +123,8 @@ jobs:
       - name: 🧪 Setup elixir
         uses: erlef/setup-elixir@v1
         with:
-          elixir-version: 1.15.5 # Define the elixir version [required]
-          otp-version: 26.1 # Define the OTP version [required]
+          elixir-version: 1.17.2 # Define the elixir version [required]
+          otp-version: 27.0.1 # Define the OTP version [required]
 
       - name: ⬇️ Install elixir dependencies
         run: mix deps.get
@@ -164,8 +164,8 @@ jobs:
       - name: 🧪 Setup elixir
         uses: erlef/setup-elixir@v1
         with:
-          elixir-version: 1.15.5 # Define the elixir version [required]
-          otp-version: 26.1 # Define the OTP version [required]
+          elixir-version: 1.17.2 # Define the elixir version [required]
+          otp-version: 27.0.1 # Define the OTP version [required]
 
       - name: ⬇️ Install elixir dependencies
         run: mix deps.get