--- a/.github/workflows/build.yml
+++ b/.github/workflows/build.yml
@@ -66,7 +66,7 @@ jobs:
         run: set -a;source oli.env && mix clean
 
       - name: 🔨 Build project
-        run: set -a; source oli.env && mix compile --warnings-as-errors
+        run: set -a; source oli.env && mix compile
 
       - name: 📋 Check Formatting
         run: mix format --check-formatted