--- a/.github/workflows/documentation.yml
+++ b/.github/workflows/documentation.yml
@@ -29,8 +29,8 @@ jobs:
       - name: 🧪 Setup Elixir
         uses: erlef/setup-elixir@v1
         with:
-          elixir-version: 1.15.5 # Define the elixir version [required]
-          otp-version: 26.1 # Define the OTP version [required]
+          elixir-version: 1.17.2 # Define the elixir version [required]
+          otp-version: 27.0.1 # Define the OTP version [required]
 
       - name: ⬇️ Install Elixir Dependencies
         run: mix deps.get