--- a/.tool-versions
+++ b/.tool-versions
@@ -1,4 +1,4 @@
-erlang 26.1
-elixir 1.15.5-otp-26
+erlang 27.0.1
+elixir 1.17.2-otp-27
 nodejs 16.14.2
 yarn 1.22.17