--- a/config/config.exs
+++ b/config/config.exs
@@ -289,14 +289,6 @@ config :tailwind,
       --output=../priv/static/css/app.css
     ),
     cd: Path.expand("../assets", __DIR__)
-  ],
-  storybook: [
-    args: ~w(
-      --config=tailwind.config.js
-      --input=css/storybook.css
-      --output=../priv/static/assets/storybook.css
-    ),
-    cd: Path.expand("../assets", __DIR__)
   ]
 
 config :ex_cldr,