--- a/config/dev.exs
+++ b/config/dev.exs
@@ -133,8 +133,7 @@ config :oli, OliWeb.Endpoint,
       "--watch-options-stdin",
       cd: Path.expand("../assets", __DIR__)
     ],
-    tailwind: {Tailwind, :install_and_run, [:default, ~w(--watch)]},
-    storybook_tailwind: {Tailwind, :install_and_run, [:storybook, ~w(--watch)]}
+    tailwind: {Tailwind, :install_and_run, [:default, ~w(--watch)]}
   ]
 
 # ## SSL Support
@@ -168,8 +167,7 @@ config :oli, OliWeb.Endpoint,
       ~r"priv/static/.*(js|css|png|jpeg|jpg|gif|svg)$",
       ~r"priv/gettext/.*(po)$",
       ~r"lib/oli_web/{live,views}/.*(ex)$",
-      ~r"lib/oli_web/templates/.*(eex)$",
-      ~r"storybook/.*(exs)$"
+      ~r"lib/oli_web/templates/.*(eex)$"
     ]
   ]