--- a/config/test.exs
+++ b/config/test.exs
@@ -2,7 +2,6 @@ import Config
 
 config :oli,
   env: :test,
-  prometheus_port: 9570,
   s3_media_bucket_name: System.get_env("TEST_S3_MEDIA_BUCKET_NAME"),
   s3_xapi_bucket_name: System.get_env("S3_XAPI_BUCKET_NAME"),
   media_url: System.get_env("TEST_MEDIA_URL"),