--- a/lib/oli/delivery/experiments/log_worker.ex
+++ b/lib/oli/delivery/experiments/log_worker.ex
@@ -49,12 +49,12 @@ defmodule Oli.Delivery.Experiments.LogWorker do
 
     correctness =
       case score do
-        0.0 ->
+        score when score in [+0.0, -0.0] ->
           0.0
 
         s ->
           case out_of do
-            0.0 -> 0.0
+            out_of when out_of in [+0.0, -0.0] -> 0.0
             o -> s / o
           end
       end