--- a/lib/oli/delivery/metrics.ex
+++ b/lib/oli/delivery/metrics.ex
@@ -1428,7 +1428,7 @@ defmodule Oli.Delivery.Metrics do
     |> Enum.into(%{}, fn {container_id, {correct, total}} ->
       proficiency =
         case total do
-          0.0 -> nil
+          total when total in [+0.0, -0.0] -> nil
           _ -> correct / total
         end