--- a/mix.exs
+++ b/mix.exs
@@ -134,7 +134,7 @@ defmodule Oli.MixProject do
       {:bamboo, "~> 2.2"},
       {:bamboo_ses, "~> 0.3.0"},
       {:bamboo_phoenix, "~> 1.0"},
-      {:base32_crockford, "~> 1.0.0"},
+      {:crockford_base32, "~> 0.7"},
       {:bcrypt_elixir, "~> 2.2"},
       {:briefly, "~> 0.5.0"},
       {:broadway, "~> 1.0.7"},