--- a/mix.exs
+++ b/mix.exs
@@ -204,7 +204,7 @@ defmodule Oli.MixProject do
       {:sched_ex, "~> 1.1"},
       {:shortuuid, "~> 2.1"},
       {:sweet_xml, "~> 0.7.1"},
-      {:tailwind, "~> 0.1.9"},
+      {:tailwind, "~> 0.2"},
       {:telemetry, "~> 0.4.1"},
       {:telemetry_poller, "~> 0.4"},
       {:telemetry_metrics, "~> 0.4"},