--- a/test/oli_web/plugs/redirect_plug_test.exs
+++ b/test/oli_web/plugs/redirect_plug_test.exs
@@ -16,7 +16,7 @@ defmodule OliWeb.Plugs.RedirectTest do
 
   test "an exception is raised when `to` or `external` isn't defined" do
     assert_raise Plug.Conn.WrapperError,
-                 ~R[Missing required to: / external: option in redirect],
+                 ~r[Missing required to: / external: option in redirect],
                  fn ->
                    call(Router, :get, "/exceptional")
                  end