Ticket [MER-3391](https://eliterate.atlassian.net/browse/MER-3391)

This PR upgraded Erlang and Elixir to `27.0.1` and `1.17.2-otp-27`, respectively.

When we changed to use an Elixir version >= 1.17.0, many dependencies threw the following warnings:
```
warning: single-quoted strings represent charlists. Use ~c"" if you indeed want a charlist or use "" instead
```

Most of the following dependencies that are part of our code haven't handled this issue yet:

```
libcluster, nodejs, poison, xml_builder, jose, dialyxir, premailex, crontab, base32_crockford, 
bamboo, hound, timex, odgn_json_pointer, tailwind, bcrypt_elixir, earmark, sweet_xml, ex_aws, 
ex_aws_s3, broadway_dashboard, uuid, makeup_eex, phoenix_storybook, mail
```

The [Base32Crockford](https://github.com/voldy/base32_crockford) dependency was replaced with [CrockfordBase32](https://github.com/xinz/crockford_base32) because the former appears to be abandoned; the last commit was 5 years ago. The new dependency was adapted to our current behavior.

The `phoenix_storybook` dependency was removed because we are not using it.

#### Dependencies

- Upgraded: `dialyxir, ex_money, poison, tailwind, xml_builder`
- Deleted: `phoenix_storybook, hound`
- Replaced: `base32_crockford` for `crockford_base32`

#### Benchmarks:

```
this branch
mix compile  216.12s user 56.98s system 220% cpu 2:03.74 total
mix compile  215.37s user 57.89s system 218% cpu 2:04.96 total

master branch
mix compile  204.78s user 57.35s system 291% cpu 1:29.89 total
mix compile  207.66s user 58.81s system 292% cpu 1:31.16 total
```



[MER-3391]: https://eliterate.atlassian.net/browse/MER-3391?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ